backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.4.2
python-decouple==3.8
python-dotenv==1.0.0
supabase==1.2.0
httpx>=0.24.0,<0.25.0
langchain>=0.1.0
openai>=1.6.1,<2.0.0
langchain-openai>=0.0.2
langgraph>=0.0.20
langsmith>=0.1.0 